SCMNEW-44: Changes in TransactionLog

+++ scm-common-log/src/main/java/ir/daneshrefah/scm/common/log/model/TransactionPayloadRequest.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Getter
 public class TransactionPayloadRequest {
     private Long transactionLogId;
-    private Integer archiveNo;
+    private Long archiveNo;
 }
